lightweight-cnn/requirements.txt
torch
torchvision
torchmetrics
requests
numpy<2.0
matplotlib
Pillow 
tensorboard 
tqdm 
hydra-core
click 